+++ markwon-linkify/src/main/java/io/noties/markwon/linkify/LinkifyPlugin.java
@@ -42,6 +42,7 @@ public class LinkifyPlugin extends AbstractMarkwonPlugin {
      * @param useCompat If true, use {@link LinkifyCompat} to handle links.
      *                  Note that the {@link LinkifyCompat} depends on androidx.core:core,
      *                  the dependency must be added on a client side explicitly.
+     * @since 4.3.0-SNAPSHOT `useCompat` argument
      */
     @NonNull
     public static LinkifyPlugin create(boolean useCompat) {
@@ -57,6 +58,7 @@ public class LinkifyPlugin extends AbstractMarkwonPlugin {
      * @param useCompat If true, use {@link LinkifyCompat} to handle links.
      *                  Note that the {@link LinkifyCompat} depends on androidx.core:core,
      *                  the dependency must be added on a client side explicitly.
+     * @since 4.3.0-SNAPSHOT `useCompat` argument
      */
     @NonNull
     public static LinkifyPlugin create(@LinkifyMask int mask, boolean useCompat) {
@@ -78,6 +80,7 @@ public class LinkifyPlugin extends AbstractMarkwonPlugin {
             @Override
             public void apply(@NonNull CorePlugin corePlugin) {
                 final LinkifyTextAddedListener listener;
+                // @since 4.3.0-SNAPSHOT
                 if (useCompat) {
                     listener = new LinkifyCompatTextAddedListener(mask);
                 } else {
@@ -137,6 +140,7 @@ public class LinkifyPlugin extends AbstractMarkwonPlugin {
         }
     }
 
+    // @since 4.3.0-SNAPSHOT
     private static class LinkifyCompatTextAddedListener extends LinkifyTextAddedListener {
 
         LinkifyCompatTextAddedListener(int mask) {